#77 refact : 메인 페이지 컴포넌트 / 훅 분리
