+++ app/_components/main/HomeContent.tsx
@@ -1,0 +1,143 @@
+// app/main/HomeContent.tsx
+import React from "react";
+import NetworkGraph from "../Network/NetworkGraph";
+import ControlPanel from "../Network/ControlPanel";
+import GroupedNodeList from "../Network/GroupedNodeList";
+import NodeConversation from "../Network/NodeConversation";
+import Video from "../Video/Video";
+import styles from "../../main/Main.module.css";
+import Header from "../common/Header";
+import Footer from "../common/Footer";
+import SharingRoom from "../sharingRoom";
+import VoiceRecorder from "../VoiceRecorder/VoiceRecorder";
+import useHomeContent from "../../_hooks/useHomeContent";
+
+const HomeContent: React.FC = () => {
+  const {
+    socket,
+    socketContext,
+    containerRef,
+    sessionId,
+    userName,
+    token,
+    isConnected,
+    nodes,
+    edges,
+    selectedNodeId,
+    handleNodeClick,
+    fitToScreen,
+    controlNodeLabel,
+    setControlNodeLabel,
+    controlNodeContent,
+    setControlNodeContent,
+    controlNodeColor,
+    setControlNodeColor,
+    roomLink,
+    isModalOpen,
+    setIsModalOpen,
+    isListOpen,
+    handleAddNode,
+    handleKeyword,
+    handleSharingRoom,
+    handleLeaveSession,
+    setLeaveSessionCallback,
+    setAction,
+    publisher,
+    subscriber,
+  } = useHomeContent();
+
+  if (!socketContext) {
+    return <p>Error: Socket context is not available.</p>;
+  }
+
+  return (
+    <div className={styles.container}>
+      <Header>MIKO</Header>
+      <div className={styles.mainContainer}>
+        <div className={styles.networkGraphContainer}>
+          <NetworkGraph
+            containerRef={containerRef}
+            selectedNodeId={selectedNodeId}
+            handleNodeClick={handleNodeClick}
+            socket={socket}
+          />
+        </div>
+
+        <div className={styles.appContainer}>
+          {isConnected ? (
+            <>
+              <div className={styles.appContainer}>
+                {sessionId && userName && token ? (
+                  <Video
+                    sessionId={sessionId}
+                    userName={userName}
+                    token={token}
+                    setLeaveSessionCallback={setLeaveSessionCallback}
+                  />
+                ) : (
+                  <p>Loading...</p>
+                )}
+              </div>
+            </>
+          ) : (
+            <p>Socket is not connected. Please check your connection.</p>
+          )}
+        </div>
+        <div className={styles.nodeConversationWrapper}>
+          <NodeConversation
+            nodes={nodes.get()}
+            selectedNodeId={selectedNodeId}
+            onNodeClick={handleNodeClick}
+          />
+        </div>
+        <div
+          className={`${styles.groupedNodeListWrapper} ${
+            isListOpen ? styles.open : ""
+          }`}
+        >
+          <GroupedNodeList
+            nodes={nodes.get()}
+            edges={edges.get()}
+            selectedNodeId={selectedNodeId}
+            onNodeClick={handleNodeClick}
+          />
+          <button className={styles.groupedNodeListButton}></button>
+        </div>
+      </div>
+      <Footer>
+        <div className={styles.footerComponents}>
+          <button onClick={handleSharingRoom}>Sharing a room</button>
+          <ControlPanel
+            newNodeLabel={controlNodeLabel}
+            newNodeContent={controlNodeContent}
+            newNodeColor={controlNodeColor}
+            setNewNodeLabel={setControlNodeLabel}
+            setNewNodeContent={setControlNodeContent}
+            setNewNodeColor={setControlNodeColor}
+            addNode={handleAddNode}
+            setAction={setAction}
+            fitToScreen={fitToScreen}
+          />
+          {sessionId && (
+            <VoiceRecorder
+              sessionId={sessionId}
+              publisher={publisher}
+              subscriber={subscriber}
+            />
+          )}
+          <button className={styles.keywordButton} onClick={handleKeyword}>
+            keyword
+          </button>
+          <button onClick={handleLeaveSession}>Leave Session</button>
+        </div>
+      </Footer>
+      <SharingRoom
+        isOpen={isModalOpen}
+        onClose={() => setIsModalOpen(false)}
+        roomLink={roomLink}
+      />
+    </div>
+  );
+};
+
+export default HomeContent;
